Add basic timing info to log_enter_exit decorator
The decorator used to wrap methods/functions has been enhanced
to report time information for tenter, texit, telapsed. Although
the tenter and texit information can be found in the log entries,
they are of less precision, only available if using the logger
(configured to report time), and there is no telapsed. This update
addresses these points.