Add SubordinateReadyRelationSuperInterface and register
Add SubordinateReadyRelationSuperInterface which is the interface
used between a principal (provides) and a subordinate (requires).
It contains a ready status (effectively, a boolean) on the
subordinate side that can be used to indicate subordinate readiness.
Also, add to the registry at "relation-subordinate-ready". This makes
is available to all principal-subordinate use cases.