Preliminary support for extensions to the library and an initial entry (file.py)
The hpctlib.ext namespace is for extensions to the library. They
are not part of the base library and are intended to be split off
from the library at some point.

The initial entry is hpctlib/ext/interfaces/file.py. Which provides
an Interface implementation for sharing/communicating/working with
files.