Update Interface to use base interface to carry out operations
Refactored so that _store_clear(), _store_get(), and _store_set()
actually carry out the operation. These, then, are overridden when
the storage target is different.

The clear(), _get(), and _set() now call the respective calls listed
above.

Add BaseInterface as base for embedded subinterfaces and support for
pointing to the base Interface.